qcow2: Work-around for proper COW synchronization
We must ensure that requests accessing the L2 tables do not intersect
with those that modify them; for example, otherwise, two requests might
want to COW the very same non-copied cluster at the same time, which
would lead to data loss.

As a work-around for proper COW synchronization, keep a global RwLock
that is write-locked around all L2 table modifications and COW
operations we do.

A proper solution would probably be to introduce an L2 table cache
(which we are going to do anyway), and then have RwLock<L2Table> objects
in there.  (We could do even better and lock sub-ranges of L2 tables,
but whether that is feasible and worth it is a different question.)

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>