file: Do not assume the end of file
Let the OS announce the end of the file by returning 0 from `preadv()`.
Do not assume it will be there.

Assuming it is there can not only lead to races, but also breaks
alignment requirements: We must read a fully aligned block, regardless
of whether it is across the EOF.  The kernel treats it as an error if we
try to be clever and align it exactly to the EOF preemptively.  Done
correctly, it will return to us the EOF offset in the buffer, we
continue to try to read from there, and it will tells us we reached the
end of the file.

Also handle EINTR and `pwritev()` returning 0.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>
(cherry picked from commit 7f82ce4fba720b16aa06ee61cb045bc26c871dda)
Signed-off-by: Hanna Czenczek <hreitz@redhat.com>