Add format::Format
This will be useful for “implicit open gating”: Image formats may
provide functionality to implicitly open dependencies like backing files
as specified by their metadata (qcow2 can do this for its backing file
(recursively) and an external data file).  We want this to be
“gateable”, i.e. allow users to intercept everything that is to be
opened implicitly, and for this we need to somehow specify the format we
want to use for the image that is about to be opened.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>