Add image open builder and gating
Using a Builder object to configure options for opening images rather
than fixed arguments to an `open_image()` function is preferable, as it
will allow adding new options to image formats in a backwards-compatible
manner.

This also unifies how images are opened behind a single trait and
pattern (`FormatDriverInstance::builder().[...].open()`).

In addition, add “implicit open gating”: When opening an image, the user
can install a gate (trait implementation) whose methods are called for
every image or storage object that is to be opened implicitly (e.g.
backing images).  The implementation is then responsible for actually
opening the image or storage object, or can refuse to do so, based on
the parameters used for opening it.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>