Radically simplify re-alignment code
Optimally realigning I/O vectors is hard and complicated, especially
when you want to do it safe Rust (i.e. properly retaining lifetimes and
ownership).

“Optimally” here means looking through the entire I/O vector, looking
for as much of an aligned subset as possible, and only creating the
absolute minimum set of bounce buffers.

I am hard-pressed to imagine situations where doing this would really be
useful.  Users of imago should ensure that the buffers they use are
aligned, so at best, all of this complicated code is dead code, which is
dangerous.

Should users however ignore alignment requirements, it seems hard to
imagine a case where they would pass an I/O vector that is partially
unaligned, but does contain fully-aligned subsets.  Note that for a
buffer to be aligned, it must both adhere to the memory and the request
alignment; in a sufficiently large request, of course you will find
portions that are aligned to the request alignment, but it seems
extremely unlikely that these would then also be aligned to the memory
alignment.

Therefore, just always create a single bounce buffer covering
everything.  This radically simplifies the code, and the previous code
wasn’t optimal either.[1]

[1] Specifically, the previous code does not look for aligned
    subsections inside of the I/O vector’s buffers, it only keeps fully
    aligned buffers intact.  While trying to remedy this, everything got
    so complicated, that I decided to commit this instead.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>
(cherry picked from commit 825025d2e8dab964d304231a7f05972173e33618)
Signed-off-by: Hanna Czenczek <hreitz@redhat.com>