Formats: Add discarding and write-zeroes
We need to be careful with anything that can discard data mappings, as
that may seriously break existing mapping information, i.e. not just
invalidate it, but make them actually dangerous to use: Data mappings
are allowed to be writable, bypassing the format layer, so if one is
invalidated, it must absolutely no longer be written to, or the results
would be undefined.  (For example, the mapping could be repurposed, and
then the data would show up somewhere else, or worse, corrupt metadata.)

Therefore, anything that can discard mappings must take a mutable
reference to `FormatAccess`/`FormatDriverInstance`.  This ensures that
any mapping returned by `get_mapping()` or `ensure_data_mapping()`
cannot be kept across discard operations, because they contain storage
object references, which are bound to the lifetime of an immutable
`FormatAccess`/`FormatDriverInstance` reference.

As an alternative, there is also the non-discarding
`FormatAccess::write_zeroes()` variant (through
`FormatDriverInstance::ensure_zero_mapping()`).

Fixes: #1
Signed-off-by: Hanna Czenczek <hreitz@redhat.com>