WrappedFormat: Implement for OwnedRwLockReadGuard
Implementing WrappedFormat for Arc<RwLock<_>>, especially tokio::RwLock,
would be great, but it’s hard.

Specifically, this would limit the lifetime of references it could
generate to that of the respective lock guard, which basically would
require carrying that lock guard everywhere while the reference needs to
be kept alive.  That requires some redesign.

A quick work-around for now is to instead implement it at least for
tokio::OwnedRwLockReadGuard<_>, i.e. keeping a constant read lock.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>