qcow2: Fix unaligned discard at the image end
When the image end is inside of a cluster, and we are asked to discard
that cluster until the image end, we can and should just discard the
whole cluster then.

The VM state (if any) lies beyond the image end, but only starting from
the first full cluster there, so discarding the last cluster of the
image has no effect on anything but the area that we want to discard.

Fixes: 0ea71defd1c5d423d41a26ef22e0b63116e60ffa
       (“Formats: Add discarding and write-zeroes”)
Signed-off-by: Hanna Czenczek <hreitz@redhat.com>