DynStorage: Fix infinite recursion
Using `<Self as DynStorage>::foo(self)` to implement `Storage` methods
on `Deref<DynStorage>` objects was a bad idea.  This did not go into the
`Deref<_>` object, and so led to an infinite recursion, ping-ponging
between those methods and the reverse `impl DynStorage for Storage`.

Fix it by calling `.as_ref()`.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>