qcow2: Fix feature name table parsing
It is possible a feature name field contains interior NUL bytes.  When
converting a feature name to a Rust string, we need to look for the
first NUL byte overall, not the first one in a tail of NUL bytes.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>
(cherry picked from commit 8a20a2ae986bba1fe46344258171d178c18a88a8)
Signed-off-by: Hanna Czenczek <hreitz@redhat.com>