qcow2: Fix large region allocation
Currently, new refblocks are allocated by having them always cover
themselves.  That makes every refblock contain one reserved entry,
basically, which prevents allocating continuous regions that are as
large as or larger than what a single refblock can cover.

For default settings, that does not really matter (64k cluster size and
16 bit refblocks means one refblock can cover 2 GB; an L1 table that
size would be sufficient for a 128 PB image), but still isn’t great.

To fix it, when allocating a new refblock, we now first try to put it
somewhere where is covered by a preceding refblock, which will allow the
whole new refblock to be empty and usable.  If there is no space before
the current refblock’s area, we self-describe it as before.

This way, when we allocate a large region, this will happen:
- We look for empty space, the current last refblock is not empty
  enough, so we allocate a new refblock.
- For this new refblock, we may still find empty space in the previous
  refblock, or not.  In the latter case, we will have it be
  self-described, in the former case, it will be empty.
- Either way, that new refblock is still not large enough to cover the
  to-be-allocated large region, so we need yet another one.  This one
  will now definitely find space in the previous (or previous previous)
  refblock, so definitely will be completely empty.
- Same goes for any consecutive refblocks, until the region eventually
  is indeed large enough.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>
(cherry picked from commit 75b1affa0c549b21c4e651d09fdc7ba4320e7380)
Signed-off-by: Hanna Czenczek <hreitz@redhat.com>