Fix clippy warning: String::len()
Clippy points out that `String::len()` returns the length in bytes, so
`x.as_bytes().len()` is the same as `x.len()`.

I want to note that I find this quite unintuitive and wonder why it
isn’t called something more explicit, like `String::byte_len()` or
`String::byte_count()`.

Anyway, a comment pointing out that this is the length in bytes should
be sufficient to prevent potential confusion.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>