CI: Pull docker image from hreitz/imago
Use hreitz/imago as the docker image source, not the MR author’s local
registry; the latter will only have an image if that author ever touched
the CI files, too.

Ideally, we would still use the local registry in case an MR does touch
the CI files, but I’m not proficient enough in gitlab magic to do so.
If an MR comes along that touches the CI files, and then needs to use
the updated docker image immediately, we’ll just need to split it.

An alternative solution could be to create the docker image in the local
registry if there currently is no such image. The problem with that is
that it would only check whether there’s an image, not whether it’s
still up-to-date, which is why I think it’s better to just pull the
image from hreitz/imago.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>
(cherry picked from commit ecb59f1c85aa670d4ceb5268535bb67bae0a18cf)
Signed-off-by: Hanna Czenczek <hreitz@redhat.com>