file: implement relaxed sync for macos
On macOS, fsync() flushes the filesystem cache to the drive, but doesn't
request the drive to flush it's internal cache to its persistent
storage. To do the latter, there's a specific fcntl: F_FULLFSYNC.

The reason why they decided to have fsync() behave this way it's because
in Apple Silicon hardware flushing the drive cache has a very
significant performance impact.

In File::sync() we were calling std::fs::File::sync_all(), which does
use F_FULLSYNC, which is safer but comes with a high cost.

Let's allow crate users on macOS opt for a more relaxed synchronization
using fsync() instead of std::fs::File::sync_all(). Expose the option
in StorageOpenOptions as relaxed_sync but gate it behind a build-time
conditional for target_os = macos.

Signed-off-by: Sergio Lopez <slp@redhat.com>