file: detect zero and discard alignment on macOS
On macOS, we're using F_PUNCHHOLE on pure_discard(), which is used for
both discard operations and writing zeroes. On APFS, this mechanism
requires that the request is aligned to the block size of the
filesystem, as otherwise the syscall returns an error (this was causing
the guest to receive I/O errors on libkrun).

As the Storage trait already supports dealing with alignment
requirements for these operations, let's use them in File. Also, let's
find out the block size for the filesystem where the backing image
resides using `libc::fstatfs`.

Signed-off-by: Sergio Lopez <slp@redhat.com>
(cherry picked from commit dd8b4daa18c163b4ff7f5f0ef207d90f6cbc8848)
Signed-off-by: Hanna Czenczek <hreitz@redhat.com>