qcow2: Add skip_cow_to_eof to do_ensure_data_map.
Our preallocation functions invoke `do_ensure_data_mapping()` beyond the
virtual disk’s boundary.  That function checks whether to do COW based
on `skip_cow` and the given `offset`/`length` combination:
- If `skip_cow` (previously `overwrite`) is false, copy everything of
  all newly allocated clusters.
- Otherwise, copy only those parts of newly allocated clusters that are
  not in the `offset`/`length` range, i.e. the head and tail around it.
  When `offset + length` is equal to the virtual disk’s end, skip the
  tail as it is beyond the virtual disk boundary.

Now, in case of preallocation, the virtual disk end is not up to date.
We will be preallocating right up to it, so there is no need to do

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>