qcow2: Have preallocate_zero() round up to clusters
We can only mark full clusters as zero clusters, so for unaligned images
we currently do not mark the last cluster as a zero cluster, instead
allocating it as a data cluster and writing zeroes to it.

It doesn’t matter what’s beyond the virtual disk end, though[1], so we
can and should just make the whole last cluster a zero cluster.  To do
so, round the preallocation length up to full clusters.

[1] QEMU can keep VM state there, but only after the last full cluster.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>